Move alias templates to parser
